fast_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
asyncpg
python-dotenv
python-multipart
jinja2
python-jose[cryptography]
passlib[bcrypt]
email-validator
aiofiles
sqlalchemy
alembic
databases[postgresql]
httpx
redis
celery
